UICIRC-382: Update structure of the circulation rules editor tests
